Refactor: extract start command
