+++ src/app-definition.type.ts
@@ -1,0 +1,4 @@
+export interface AppDefinition {
+  name: string;
+  script: string;
+}
